pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "secondbrain-voice-gateway"
version = "0.1.0"
description = "Self-hosted Alexa voice gateway for SecondBrain, Home Assistant, and Docker operations."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary - Feberdin internal use unless you add a LICENSE file." }
authors = [
  { name = "Feberdin" }
]
dependencies = [
  "cryptography>=45.0.0,<46.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "httpx>=0.28.0,<0.29.0",
  "pydantic>=2.11.0,<3.0.0",
  "pydantic-settings>=2.9.0,<3.0.0",
  "PyYAML>=6.0.2,<7.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=1.0.0,<2.0.0"
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
asyncio_mode = "auto"

